requirements.txt
django==2.2.2
psycopg2==2.8.3
Pillow==6.0.0
pytz==2019.1
django-phone-field==1.7.1
